+++ src/test/java/com/alibaba/json/bvt/serializer/date/DateTest_ISO8601_ThreeLetterISO8601TimeZone.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.date;
 
 import java.util.Calendar;
 
@@ -8,10 +8,10 @@ import org.junit.Assert;
 
 import com.alibaba.fastjson.JSON;
 
-public class DateTest_ISO8601_OneLetterISO8601TimeZone extends TestCase {
+public class DateTest_ISO8601_ThreeLetterISO8601TimeZone extends TestCase {
 
     public void f_test_date() throws Exception {
-        String text = "{\"gmtCreate\":\"2014-08-21T09:51:36.2+08\"}";
+        String text = "{\"gmtCreate\":\"2014-08-21T09:51:36.2+08:00\"}";
         
         Calendar date = JSON.parseObject(text, VO.class).getGmtCreate();
         Assert.assertNotNull(date);
@@ -20,7 +20,7 @@ public class DateTest_ISO8601_OneLetterISO8601TimeZone extends TestCase {
     }
     
     public void test_date_9() throws Exception {
-        String text = "{\"gmtCreate\":\"2014-08-21T09:51:36.235+09\"}";
+        String text = "{\"gmtCreate\":\"2014-08-21T09:51:36.235+09:00\"}";
         
         Calendar date = JSON.parseObject(text, VO.class).getGmtCreate();
         Assert.assertNotNull(date);
