+++ src/test/java/com/alibaba/json/bvt/serializer/date/DateTest_ISO8601_UTCTime.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.date;
 
 import java.util.Calendar;
